requirements.txt
langchain>=0.0.1
langgraph>=0.0.20
langchain-groq>=0.0.1
streamlit>=1.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
hypothesis>=6.75.3
pytest>=7.4.0
diskcache>=5.6.1
pylint>=2.17.0